requirements-minimal.txt
# Minimal requirements for baseball data ingestion
# Python 3.9 compatible

# Database
supabase>=1.0.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0

# HTTP clients
requests>=2.31.0
httpx>=0.25.0

# Environment
python-dotenv>=1.0.0

# Baseball Data Sources
pybaseball>=2.2.1
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Task Scheduling
schedule>=1.2.0

# Utilities
tqdm>=4.66.0
python-dateutil>=2.8.0